RevokeMigratorRoleCommandBaseHandler
