+++ src/Octoshift/Commands/RevokeMigratorRoleCommandBaseHandler.cs
@@ -1,0 +1,66 @@
+using System;
+using System.CommandLine;
+using System.Threading.Tasks;
+using OctoshiftCLI.Contracts;
+using OctoshiftCLI.Extensions;
+
+namespace OctoshiftCLI.Commands;
+
+public class RevokeMigratorRoleCommandBaseHandler
+{
+    private readonly OctoLogger _log;
+    private readonly ITargetGithubApiFactory _githubApiFactory;
+
+    public RevokeMigratorRoleCommandBaseHandler(OctoLogger log, ITargetGithubApiFactory githubApiFactory)
+    {
+        _log = log;
+        _githubApiFactory = githubApiFactory;
+    }
+
+    public async Task Handle(RevokeMigratorRoleArgs args)
+    {
+        if (args is null)
+        {
+            throw new ArgumentNullException(nameof(args));
+        }
+
+        _log.Verbose = args.Verbose;
+
+        _log.LogInformation("Granting migrator role ...");
+        _log.LogInformation($"GITHUB ORG: {args.GithubOrg}");
+        _log.LogInformation($"ACTOR: {args.Actor}");
+
+        if (args.GithubPat is not null)
+        {
+            _log.LogInformation($"GITHUB PAT: ***");
+        }
+
+        args.ActorType = args.ActorType?.ToUpper();
+        _log.LogInformation($"ACTOR TYPE: {args.ActorType}");
+
+        args.ActorType = args.ActorType.ToUpper();
+
+        if (args.ActorType is "TEAM" or "USER")
+        {
+            _log.LogInformation("Actor type is valid...");
+        }
+        else
+        {
+            _log.LogError("Actor type must be either TEAM or USER.");
+            return;
+        }
+
+        var githubApi = _githubApiFactory.Create(targetPersonalAccessToken: args.GithubPat);
+        var githubOrgId = await githubApi.GetOrganizationId(args.GithubOrg);
+        var success = await githubApi.RevokeMigratorRole(githubOrgId, args.Actor, args.ActorType);
+
+        if (success)
+        {
+            _log.LogSuccess($"Migrator role successfully revoked for the {args.ActorType} \"{args.Actor}\"");
+        }
+        else
+        {
+            _log.LogError($"Migrator role couldn't be revoked for the {args.ActorType} \"{args.Actor}\"");
+        }
+    }
+}
