+++ src/OctoshiftCLI.Tests/Octoshift/Commands/RevokeMigratorRoleCommandBaseTests.cs
@@ -13,7 +13,7 @@ public class RevokeMigratorRoleCommandBaseTests
     private readonly Mock<ITargetGithubApiFactory> _mockGithubApiFactory = new();
     private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
 
-    private readonly RevokeMigratorRoleCommandBase _command;
+    private readonly RevokeMigratorRoleCommandBaseHandler _command;
 
     private const string GITHUB_ORG = "FooOrg";
     private const string ACTOR = "foo-actor";
@@ -21,7 +21,7 @@ public class RevokeMigratorRoleCommandBaseTests
 
     public RevokeMigratorRoleCommandBaseTests()
     {
-        _command = new RevokeMigratorRoleCommandBase(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
+        _command = new RevokeMigratorRoleCommandBaseHandler(_mockOctoLogger.Object, _mockGithubApiFactory.Object);
     }
 
     [Fact]
