+++ src/ado2gh/Commands/RevokeMigratorRoleCommand.cs
@@ -14,6 +14,6 @@ public sealed class RevokeMigratorRoleCommand : RevokeMigratorRoleCommandBase
         Description += $"Note: Expects GH_PAT env variable or --{GithubPat.GetLogFriendlyName()} option to be set.";
 
         AddOptions();
-        Handler = CommandHandler.Create<RevokeMigratorRoleArgs>(Handle);
+        Handler = CommandHandler.Create<RevokeMigratorRoleArgs>(BaseHandler.Handle);
     }
 }
